requirements.txt
openai
polib
python-dotenv